+++ examples/customization-top-menu-layout/src/components/sider/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { useTitle } from "@pankod/refine-core";
-import { Menu, useMenu } from "@pankod/refine-antd";
+import { useTitle, useMenu } from "@pankod/refine-core";
+import { Menu } from "@pankod/refine-antd";
 import routerProvider from "@pankod/refine-react-router-v6";
 
 const { Link } = routerProvider;