+++ examples/mern-dashboard-server/mongodb/models/user.js
@@ -1,0 +1,12 @@
+import mongoose from "mongoose";
+
+const UserSchema = new mongoose.Schema({
+    name: { type: String, required: true },
+    email: { type: String, required: true },
+    avatar: { type: String, required: true },
+    allProperties: [{ type: mongoose.Schema.Types.ObjectId, ref: "Property" }],
+});
+
+const userModel = mongoose.model("User", UserSchema);
+
+export default userModel;