+++ packages/inferencer/src/field-inferencers/number.ts
@@ -1,9 +1,15 @@
 import { FieldInferencer } from "@/types";
 
 export const numberInfer: FieldInferencer = (key, value) => {
+    const isNonEmptyString = typeof value === "string" && value.length > 0;
+    const isNotNaN = !isNaN(value as number);
+
+    const isNumericString = isNonEmptyString && isNotNaN;
     const isNumber = typeof value === "number";
 
-    if (isNumber) {
+    const isNumeric = isNumericString || isNumber;
+
+    if (isNumeric) {
         return {
             key,
             type: "number",